In progress 24.03

+++ shop-api/src/main/java/com/task/shop/api/client/ProductClient.java
@@ -21,8 +21,8 @@ public class ProductClient {
         return restTemplate.getForEntity(uriProduct + id.toString(), ProductDto.class);
     }
 
-    public ResponseEntity<ProductListDto> getAllProduct() {
-        return restTemplate.getForEntity(uriProduct + "/all", ProductListDto.class);
+    public ResponseEntity<ProductDtoList> getAllProduct() {
+        return restTemplate.getForEntity(uriProduct + "/all", ProductDtoList.class);
     }
 
 
@@ -30,16 +30,16 @@ public class ProductClient {
         return ResponseEntity.ok(restTemplate.patchForObject(uriProduct + productId, productDto, ProductDto.class));
     }
 
-    public ResponseEntity<DiscountDto> addDiscount(String productId, DiscountDto discountDto) {
+    public ResponseEntity<DiscountDto> addDiscount(DiscountDto discountDto) {
         return ResponseEntity.ok(restTemplate.postForObject(
-                uriProduct + productId + "/discount",
+                uriProduct + "discount",
                 discountDto,
                 DiscountDto.class));
     }
 
-    public DiscountDto updateDiscount(Long productId, Long discId, DiscountDto discountDto) {
+    public DiscountDto updateDiscount(Long discId, DiscountDto discountDto) {
         return ResponseEntity.ok(restTemplate.patchForObject(
-                uriProduct + productId + "/discount/" + discId,
+                uriProduct + "discount/" + discId,
                 discountDto,
                 DiscountDto.class)).getBody();
     }
@@ -54,16 +54,24 @@ public class ProductClient {
                 .getBody();
     }
 
-    public CommentDto getComment(Long productId, CommentDto commentId) {
+    public CommentDto getComment(Long productId, Long commentId) {
         return restTemplate.getForEntity(uriProduct + productId + "/comment/" + commentId, CommentDto.class).getBody();
     }
 
     public void deleteCommentByUser(Long productId, Long commentId, Authentication authentication) {
         UserDto user = userClient.getUserByName(authentication.getName());
-        restTemplate.delete(uriProduct + productId + "/comment/" + commentId + "/user/" + user.getId());
+        CommentDto commentDto = getComment(productId, commentId);
+        if (user.getId().equals(commentDto.getUserId()) || user.getRole().equals("ADMIN")) {
+            restTemplate.delete(uriProduct + productId + "/comment/" + commentId + "/user/" + user.getId());
+        } else {
+            throw new RuntimeException();
+        }
     }
 
-    public CommentDto updateCommentByUser(Long productId, Long commentId, CommentDto commentDto, Authentication authentication) {
+    public CommentDto updateCommentByUser(Long productId,
+                                          Long commentId,
+                                          CommentDto commentDto,
+                                          Authentication authentication) {
         UserDto user = userClient.getUserByName(authentication.getName());
         commentDto.setUserId(user.getId());
         return restTemplate.patchForObject(
@@ -77,5 +85,10 @@ public class ProductClient {
                 uriProduct + productId + "/rating",
                 productDto,
                 ProductDto.class);
+    }
+
+    public ProductDtoList addDiscountForProducts(ProductDtoList productDtoList) {
+        return restTemplate.postForEntity(uriProduct + "/discount/products",
+                productDtoList, ProductDtoList.class).getBody();
     }
 }
